Main Visual Studio Project for the system
Hey guys, so I created a project for our system which contains all the forms we need that correspond to our scope. I added the connection string and the database connection in each form. Please do not change it. All you have to do is to navigate to Solution explorer and select the form you have to design and code for. You right click a form, select view designer and you can move it from there.